Began working on the main UI grid layout
As the title states, began construction of the basic frame/skeleton of the Sputnik UI. At this point, a few different UI elements are still carrying default values (hence why SfTreeView is populated with a list of vegetables) or placeholder texts.

In further steps, this will be rectified and many of these options will be linked to an initial dataset, which has begun construction as StartupSatellites.json.